Research/BAMM-Unimumo/UniMuMo/requirements.txt
av==12.0.0
Cython==3.0.7
dtw-python==1.3.0
einops==0.7.0
flashy==0.0.2
julius==0.2.7
librosa==0.10.1
matplotlib==3.3.4
numpy==1.23.5
num2words
omegaconf==2.3.0
pandas==1.5.3
pytorch-lightning==1.9.1
sentencepiece
spacy==3.7.2
tensorboard==2.12.3
transformers==4.35.2
openai
